refactor: DOM조작과 순수HTML 생성 책임 분리
